+++ R/Authenticate.R
@@ -6,11 +6,12 @@
 #'   \code{Authenticate} is the first step of the \code{Authenticate}, \code{\link{Collect}} and \code{\link{Create}}
 #'   workflow.
 #'
-#'   Refer to \code{\link{Authenticate.twitter}}, \code{\link{Authenticate.youtube}} and
-#'   \code{\link{Authenticate.reddit}}, \code{\link{Authenticate.web}} for parameters and usage.
+#'   Refer to \code{\link{Authenticate.mastodon}}, \code{\link{Authenticate.youtube}},
+#'    \code{\link{Authenticate.reddit}}, \code{\link{Authenticate.twitter}} and
+#'    \code{\link{Authenticate.web}} for parameters and usage.
 #'
 #' @param socialmedia Character string. Identifier for social media API to authenticate with. Supported social media are
-#'   \code{"twitter"}, \code{"youtube"}, \code{"reddit"} and \code{"web"}.
+#'   \code{"mastodon"}, \code{"youtube"}, \code{"reddit"}, \code{"twitter"} and \code{"web"}.
 #' @param ... Optional parameters to pass to functions providied by supporting R packages that are used for social media
 #'   API access.
 #'