+++ R/Collect.listing.reddit.R
@@ -14,7 +14,7 @@
 #'   \code{"all"}.
 #' @param max Numeric vector. Maximum number of threads in listing to return. Default is \code{25}.
 #' @param waitTime Numeric vector. Time range in seconds to select random wait from in-between url collection requests.
-#'   Minimum is 3 seconds. Default is \code{c(3, 5)} for a wait time chosen from between 3 and 5 seconds.
+#'   Minimum is 3 seconds. Default is \code{c(6, 8)} for a wait time chosen from between 6 and 8 seconds.
 #' @param ua Character string. Override User-Agent string to use in Reddit thread requests. Default is
 #'   \code{option("HTTPUserAgent")} value as set by vosonSML.
 #' @param writeToFile Logical. Write collected data to file. Default is \code{FALSE}.
@@ -40,7 +40,7 @@ Collect.listing.reddit <-
            sort = "hot",
            period = "all",
            max = 25,
-           waitTime = c(3, 5),
+           waitTime = c(6, 8),
            ua = getOption("HTTPUserAgent"),
            writeToFile = FALSE,
            verbose = FALSE,